Resolve "implement interpreter of Fildex"

1. refactor `core`, set `UFS3` to variable, ‘cause the file is immutalbe, and will changed to a new one when `write`
2. add layout for `filler` and `fildex`
3. adt refactor

Closes #11

See merge request !29

+++ interpreter/src/main/scala/ufs3/block.scala
@@ -63,7 +63,7 @@ trait BlockInterpreter extends Block.Handler[Kleisli[IO, Config, ?]] {
     Kleisli { config ⇒
       IO {
         val size = if (data.position() == 0) data.limit() else data.flip().limit()
-        blockFile.write(data, size.toLong)
+        blockFile.write(data, size)
       }
     }
